PANA-8612: Split composition package files to one declaration per file
