+++ features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/composition/CapturePipelineSelector.kt
@@ -1,0 +1,17 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2016-Present Datadog, Inc.
+ */
+
+package com.datadog.android.sessionreplay.internal.composition
+
+import com.datadog.android.sessionreplay.internal.recorder.Recorder
+
+internal class CapturePipelineSelector(
+    private val compositionEnabled: Boolean,
+    private val legacyFactory: () -> Recorder,
+    private val compositionFactory: () -> Recorder
+) {
+    fun create(): Recorder = if (compositionEnabled) compositionFactory() else legacyFactory()
+}
