+++ features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/composition/CaptureWorkToken.kt
@@ -6,8 +6,6 @@
 
 package com.datadog.android.sessionreplay.internal.composition
 
-import java.util.Collections
-import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.atomic.AtomicBoolean
 
 /** Callback identity allocated from, and invalidated with, one generation scope. */
@@ -27,37 +25,5 @@ internal class CaptureWorkToken internal constructor(
 
     internal fun invalidate() {
         valid.set(false)
-    }
-}
-
-internal class GenerationWorkRegistry {
-    // Collections.newSetFromMap throws IllegalArgumentException only for a non-empty map;
-    // these maps are freshly constructed and always empty.
-    @Suppress("UnsafeThirdPartyFunctionCall")
-    private val workTokens = Collections.newSetFromMap(ConcurrentHashMap<CaptureWorkToken, Boolean>())
-
-    @Suppress("UnsafeThirdPartyFunctionCall")
-    private val trackedWork = Collections.newSetFromMap(ConcurrentHashMap<CancellableCaptureWork, Boolean>())
-
-    fun createToken(context: CaptureGenerationContext): CaptureWorkToken =
-        CaptureWorkToken(context, this).also(workTokens::add)
-
-    fun track(work: CancellableCaptureWork) {
-        trackedWork += work
-    }
-
-    // ConcurrentHashMap.remove throws NPE only for a null key; `work` is a non-null Kotlin type.
-    @Suppress("UnsafeThirdPartyFunctionCall")
-    fun release(work: CancellableCaptureWork): Boolean = trackedWork.remove(work)
-
-    fun release(token: CaptureWorkToken) {
-        workTokens -= token
-    }
-
-    fun invalidateAll() {
-        workTokens.forEach(CaptureWorkToken::invalidate)
-        workTokens.clear()
-        trackedWork.forEach(CancellableCaptureWork::cancel)
-        trackedWork.clear()
     }
 }
