+++ features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/composition/CompletedSnapshotConsumer.kt
@@ -1,0 +1,17 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2016-Present Datadog, Inc.
+ */
+
+package com.datadog.android.sessionreplay.internal.composition
+
+internal data class CompletedSnapshotCapture(
+    val generation: CaptureGenerationContext,
+    val snapshot: CapturedFullSnapshot
+)
+
+/** Receives only complete snapshots; traversal and enrichment never enter the downstream queue. */
+internal fun interface CompletedSnapshotConsumer {
+    fun consume(capture: CompletedSnapshotCapture)
+}
