+++ features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/composition/DefaultCapturedTreeValidator.kt
@@ -1,0 +1,18 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2016-Present Datadog, Inc.
+ */
+
+package com.datadog.android.sessionreplay.internal.composition
+
+internal class DefaultCapturedTreeValidator : CapturedTreeValidator {
+
+    override fun validate(snapshot: CapturedFullSnapshot): CaptureValidationResult =
+        CapturedSnapshotValidation(snapshot).validate()
+
+    override fun validate(
+        mutation: CapturedMutationSet,
+        base: CapturedFullSnapshot
+    ): CaptureValidationResult = CapturedMutationValidation(mutation, base).validate()
+}
